Move package entropy import patches to their own separate file
In preparation to disable them in the future.
