+++ src/pyodide/internal/topLevelEntropy/entropy_import_context_packages.py.d.ts
@@ -1,0 +1,6 @@
+// Copyright (c) 2026 Cloudflare, Inc.
+// Licensed under the Apache 2.0 license found in the LICENSE file or at:
+//     https://opensource.org/licenses/Apache-2.0
+
+declare const buf: ArrayBuffer;
+export default buf;
